partitions/efi: Add 'gpt_sector' kernel cmdline parameter
The gpt_sector=<sector> causes the GPT partition search to look at the
specified sector for a valid GPT header if the GPT is not found at the
beginning or the end of block device.

In particular this is needed for NVIDIA Tegra consumer-grade Android
devices in order to make them usable with the upstream kernel because
these devices use a proprietary / closed-source partition table format
for the EMMC and it's impossible to change the partition's format. Luckily
there is a GPT table in addition to the proprietary table, which is placed
in uncommon location of the EMMC storage and bootloader passes the
location to kernel using "gpt gpt_sector=<sector>" cmdline parameters.

This patch is based on the original work done by Colin Cross for the
downstream Android kernel.

Cc: Colin Cross <ccross@android.com>
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>